[projectfirma/#1789] Allow Users to Add/Edit Attachments some code adjustments
